zettel-eval/pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "zettel-eval"
version = "0.1.0"
description = "Dataset ingestion and retrieval evaluation scaffolding for zettel-style corpora."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
  "pyyaml>=6.0.2",
]

[project.scripts]
zettel-eval = "zettel_eval.cli:main"

[tool.setuptools.packages.find]
where = ["src"]
include = ["zettel_eval*"]
